+++ src/bot/.env.example
@@ -6,6 +6,9 @@ NETWORK=mainnet (mainnet|testnet)
 API_USER=UserName
 API_PASS=UserPassword
 
+SECRET_PHRASE=(generate wallets from 12 words)
+BTC_DELIVERY_PATH=(delivery path to wallet - default m/44'/0'/0'/0/0)
+ETH_DELIVERY_PATH=(delivery path to wallet - default m/44'/60'/0'/0/0)
 BLOCKCYPHER_API_TOKEN=
 
 WEB3_TESTNET_PROVIDER=https://mainnet.infura.io/v3/5ffc47f65c4042ce847ef66a3fa70d4c
