+++ src/front/shared/redux/actions/ghost.js
@@ -15,6 +15,9 @@ import config from 'app-config'
 const bitcore = require('ghost-bitcore-lib');
 import { localisePrefix } from 'helpers/locale'
 
+import { mnemonic as mnemonicUtils } from '../../../../common/utils/mnemonic'
+
+
 
 const hasAdminFee = (config
   && config.opts
@@ -26,7 +29,7 @@ const hasAdminFee = (config
 ) ? config.opts.fee.ghost : false
 
 const getRandomMnemonicWords = () => bip39.generateMnemonic()
-const validateMnemonicWords = (mnemonic) => bip39.validateMnemonic(mnemonic)
+const validateMnemonicWords = (mnemonic) => bip39.validateMnemonic(mnemonicUtils.convertMnemonicToValid(mnemonic))
 
 
 const sweepToMnemonic = (mnemonic, path) => {
@@ -75,26 +78,9 @@ const getSweepAddress = () => {
 }
 
 const getWalletByWords = (mnemonic, walletNumber = 0, path) => {
-  const seed = bip39.mnemonicToSeedSync(mnemonic)
-  const root = bip32.fromSeed(seed, ghost.network)
-  const node = root.derivePath((path) || `m/44'/0'/0'/0/${walletNumber}`)
-
-  const account = bitcoin.payments.p2pkh({
-    pubkey: node.publicKey,
-    network: ghost.network,
-  })
-
-  return {
-    mnemonic,
-    address: account.address,
-    publicKey: node.publicKey.toString('Hex'),
-    WIF: node.toWIF(),
-    node,
-    account,
-  }
+  return mnemonicUtils.getGhostWallet(ghost.network, mnemonic, walletNumber, path)
 }
 
-window.getWalletByWords = getWalletByWords
 
 const auth = (privateKey) => {
   if (privateKey) {
